broca/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
telethon>=1.28.5
flask>=2.3.3
aiosqlite>=0.19.0
letta_client>=0.1.0  # Custom package for Letta API

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.7.0
flake8>=6.1.0

# Optional dependencies
python-dateutil>=2.8.2
typing-extensions>=4.7.1